fix(GamblerTest) : 누락 gambler Test 추가
